Added linked_binary_tree module This data structure is similar to the binary tree except it has a list for data.
When data is inserted, it will be added to the list

This is useful for a completed records data structure. We can use the malloced ptr as a key. Probably better for a finder function because of LKF_APPROX
This structure will help us free the most recent generation of that malloced pointer