Changed definition of a record Changed record to have more information and to require less malloc calls than the previous (which used two malloced pointers, this one uses only one, for the struct itself). Used a union for malloc and free records to save space
One weakness of this new record struct is it's a very long path to get to a malloc or free record, for example, for a free record:

record->sub_record.free_record

This promotes ugly code. Will fix at a later point if a better way is thought of

Also changed the methods to take in less arguments. This requires the caller (lkmalloc.c) to create an LK_RECORD, but also balances the workload